requirements.txt
certifi==2018.4.16
Django==2.0.5
django-bootstrap4==0.0.6
django-taggit==0.22.2
elasticsearch==6.3.0
elasticsearch-dsl==6.2.1
ipaddress==1.0.22
Markdown==2.6.11
mysqlclient==1.3.13
Pillow==5.2.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
urllib3==1.23
